+++ src/game_q3a.cpp
@@ -37,7 +37,7 @@ enum {
 	G_TESTPRINTFLOAT        = 113
 };
 
-const char* Q3A_eng_msg_names(int cmd) {
+const char* Q3A_eng_msg_names(intptr_t cmd) {
 	switch(cmd) {
 		GEN_CASE(G_PRINT);
 		GEN_CASE(G_ERROR);
@@ -241,7 +241,7 @@ const char* Q3A_eng_msg_names(int cmd) {
 	}
 }
 
-const char* Q3A_mod_msg_names(int cmd) {
+const char* Q3A_mod_msg_names(intptr_t cmd) {
 	switch(cmd) {
 		GEN_CASE(GAME_INIT);
 		GEN_CASE(GAME_SHUTDOWN);
@@ -260,7 +260,7 @@ const char* Q3A_mod_msg_names(int cmd) {
 }
 
 
-bool Q3A_is_mod_trace_msg(int cmd) {
+bool Q3A_is_mod_trace_msg(intptr_t cmd) {
 	switch (cmd) {
 		case GAME_CLIENT_THINK:
 		case GAME_RUN_FRAME:
